Ability to configure a url signer
